Fix bugs: decimal value, enhanced input parse

+++ cpp/cpp09/ex00/BitcoinExchange.hpp
@@ -7,13 +7,14 @@
 #include <map>
 #include <cstdlib>
 #include <iomanip>
+#include <cmath>
 
 class BitcoinExchange {
 private:
 	std::map<std::string, double>	_bitcoinRates;
 	
 	void	loadBtcRates();
-	void	findBtcRates( std::string &key, int &value ) const;
+	void	findBtcRates( std::string &key, double &value ) const;
 
 public:
 	// canonical form
